+++ pubspec.yaml
@@ -32,11 +32,13 @@ dependencies:
     sdk: flutter
 
 
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   firebase_auth: ^4.15.3
   firebase_core: ^2.24.2
+  animated_splash_screen: ^1.3.0
 
 dev_dependencies:
   flutter_test:
